pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-redis>=5.4.0",
    "django>=5.1.4",
    "djangorestframework-simplejwt>=5.3.1",
    "djangorestframework>=3.15.2",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
]
